bpf: Add tuple space search table.
This implementation builds on the lib/tuple_table to expose it via BPF.

The full_key for this map is split into two, and interpreted as a key
part and a mask part. The mask defines which bits in the key are
relevant to select this particular element. When inserting, the full_key
(ie, including both key and mask) must be provided; when performing a
lookup, only the key portion is required to find the target element.

Signed-off-by: Joe Stringer <joe@ovn.org>