openvswitch: Use pcpu array for masked key lookup
Previously, when OVS performed a lookup, it declared a sw_flow_key on the
stack for masked key lookup, then for each mask, copy the unmasked key
into the stack copy, applied the mask, and performed the lookup.

However, upcoming patches will abstract out the masked lookup structure
to allow arbitrary keys and key lengths. Instead of declaring a large
buffer that is big enough for abritrary keys on the stack, allocate the
structure at table initialization time with enough space for each cpu to
have its own space to determine masked keys for lookup.

Signed-off-by: Joe Stringer <joe@ovn.org>