lib: Introduce tuple table.
Introduce lib/tuple_table which provides an implementation of
unprioritized tuple space search based on the implementation from OVS.

Elements are inserted using a key and mask. There is no ordering defined
between elements that overlap; it's up to the user to deal with the
consequences of this, by, for instance, ensuring that overlapping masks
are never inserted into the table.

Signed-off-by: Joe Stringer <joe@ovn.org>