bpf: Add helper to retrieve sock information
This patch adds a new BPF helper function, sk_lookup() which allows BPF
programs to find out if there is a socket listening on this host, and
provide some information about this socket, such as whether it is
initiating an outbound connection or listening for new inbound
connections, and whether there is any mark associated with the socket.

The signature explicitly avoids requiring an SKB, so that it may be used
from XDP to perform early drop if the traffic is not intended for the
local host.

By way of example, the following pseudocode would filter inbound
connections at XDP if there is no corresponding service listening for
the traffic:

  struct bpf_sock_tuple tuple;
  struct bpf_sock_info info;
  int err;

  populate_tuple(ctx, &tuple); // Extract the 5tuple from the packet
  err = bpf_sk_lookup_(&tuple, sizeof tuple, &info, sizeof info, 0);
  if (err == -ENOENT) {
    // Couldn't find a socket listening for this traffic. Drop.
    return XDP_DROP;
  }
  return XDP_PASS;

Signed-off-by: Joe Stringer <joe@wand.net.nz>