bpf: Add reference tracking to verifier
Allow helper functions to acquire a reference and return it into a
register. Specific pointer types such as the PTR_TO_SOCKET will
implicitly represent such a reference. The verifier must ensure that
these references are released exactly once in each path through the
program.

To achieve this, this commit ensures that registers holding values of
with reference type are not cleared unless they are otherwise copied to
another register or the stack, and that at the end of the program there
are no remaining references which have not been released. Given that
there's only one right now, there's no pairing up of the corresponding
lookup()/release() functions but this should be trivial to add on top
when necessary.

Signed-off-by: Joe Stringer <joe@wand.net.nz>