bpf: Add transparent socket support
Add support for TPROXY via a new bpf helper, bpf_skb_set_socket().

This helper requires the BPF program to discover the socket via a call
to bpf_sk_lookup_xxx(), then pass this socket to the new helper with an
explicit statement that the transparent proxy feature is in use. The
socket must have the transparent option enabled out-of-band, and must
not be currently being destroyed. If all of these conditions hold, the
socket will be assigned to the skb to allow delivery to the socket.

Signed-off-by: Joe Stringer <joe@wand.net.nz>