bpf: Support locating specific socket types
Add socket lookup flags to all socket lookup functions to allow locating
particular types of sockets:

  BPF_F_SKL_NO_EST:    Do not return established sockets
  BPF_F_SKL_NO_LISTEN: Do not return listener sockets

These flags are chosen to retain the existing behaviour for flags=0 for
backward compatibility, in which multiple lookups occur to return either
the established socket (if found) or fall back to the listen socket.

This will be useful for cases where BPF programs only intend to attempt
to find one kind of socket before modifying the lookup tuple to locate
another type of socket, such as for implementing transparent proxy.

Signed-off-by: Joe Stringer <joe@wand.net.nz>