bpf: Add socket assign support
Add support for TPROXY via a new bpf helper, bpf_sk_assign().

This helper requires the BPF program to discover the socket via a call
to bpf_sk*_lookup_*(), then pass this socket to the new helper. The
helper takes its own reference to the socket in addition to any existing
reference that may or may not currently be obtained for the duration of
BPF processing. For the destination socket to receive the traffic, the
traffic must be routed towards that socket via local route, and the
socket must have the transparent option enabled out-of-band. If all of
these conditions hold, the socket will be assigned to the skb to allow
delivery to the socket.

Signed-off-by: Joe Stringer <joe@wand.net.nz>