bpf: Add pressure signal support to LRU updates
In order to track map churn accurately and assess the effectiveness of
the map in achieving the intended goals of the BPF program given the
access patterns, it is useful to be able to track when updates to the
LRU maps trigger the LRU property. This can be an initial signal of
"some" load on the system, which can then be tracked or combined with
other signals to assist management plane scheduling decisions in
distributed systems.

This patch adds BPF_F_PRESSURE for this purpose. If the flag is not set,
then map updates will continue to operate as they have been, where even
if the LRU property is triggered the function will return 0. However, if
the BPF program wishes to take additional action when the LRU map is
under pressure, it can pass BPF_F_PRESSURE to the map update functions,
and the BPF API will return -EINPROGRESS to indicate that an old entry
was removed and the new entry has taken its place in the map.

Signed-off-by: Joe Stringer <joe@isovalent.com>